fixed generate_manual_data to work with radiation script

+++ agml/synthetic/manual.py
@@ -14,6 +14,7 @@
 
 import os
 import re
+import stat
 import sys
 import subprocess as sp
 
@@ -101,7 +102,7 @@ def generate_manual_data(code = None,
                         f"could not be found. Check that Helios is installed there.")
 
         # Check that the necessary files exist.
-        if not os.path.exists(os.path.join(project_path, 'main.cpp')):
+        if not os.path.exists(os.path.join(project_path, 'main.cpp')) and not os.path.exists(os.path.join(project_path, 'generate.cpp')):
             raise ValueError(f"Could not find `main.cpp` file at {project_path}.")
         if not os.path.exists(os.path.join(project_path, 'CMakeLists.txt')):
             raise ValueError(f"Could not find `CMakeLists.txt` file at {project_path}.")
@@ -113,8 +114,12 @@ def generate_manual_data(code = None,
         # Check if the C++ file has changed.
         with open(os.path.join(SUPER_BASE_DIR, '.last_manual_compilation_cpp.cpp'), 'r') as f:
             legacy_cpp = f.read()
-        with open(os.path.join(project_path, 'main.cpp'), 'r') as f:
-            current_cpp = f.read()
+        try:
+            with open(os.path.join(project_path, 'main.cpp'), 'r') as f:
+                current_cpp = f.read()
+        except FileNotFoundError:
+            with open(os.path.join(project_path, 'generate.cpp'), 'r') as f:
+                current_cpp = f.read()
         if legacy_cpp == current_cpp:
             cpp_same = True
 
@@ -135,7 +140,8 @@ def generate_manual_data(code = None,
                       f'-DCMAKE_BUILD_TYPE={cmake_build_type}']
         make_args = ['cmake', '--build', '.']
         log_file = os.path.join(SUPER_BASE_DIR, '.last_helios_manual_compilation.log')
-        os.unlink(log_file)
+        if os.path.exists(log_file):
+            os.unlink(log_file)
 
         # Create the build directory again, since it has been moved.
         os.makedirs(helios_build, exist_ok = True)
@@ -186,8 +192,13 @@ def generate_manual_data(code = None,
 
     # Run the generation.
     executable = os.path.join(project_path, 'build', project_name)
+    os.chmod(executable, stat.S_IRUSR | stat.S_IWUSR | stat.S_IXUSR | stat.S_IRGRP | stat.S_IXGRP |
+             stat.S_IROTH | stat.S_IXOTH) # add necessary permissions to executable
+    env = os.environ.copy()
+    if os.path.exists('/usr/lib/wsl/lib'): # if using WSL, this is mandatory to use radiation plugin
+        env['LD_LIBRARY_PATH'] += ':/usr/lib/wsl/lib'
     process = sp.Popen([executable], stdout = sp.PIPE, stderr = sp.STDOUT,
-                       cwd = os.path.join(project_path, 'build'),
+                       cwd = os.path.join(project_path, 'build'), env=env,
                        universal_newlines = True, )
     for line in iter(process.stdout.readline, ""):
         sys.stdout.write(line)
@@ -196,9 +207,15 @@ def generate_manual_data(code = None,
 
     # Save the existing files for comparison in future runs.
     with open(os.path.join(SUPER_BASE_DIR, '.last_manual_compilation_cpp.cpp'), 'w') as f:
-        f.write(code)
+        if code is not None and not os.path.exists(code):
+            f.write(code)
+        else:
+            f.write(current_cpp)
     with open(os.path.join(SUPER_BASE_DIR, '.last_manual_compilation_cmake.txt'), 'w') as f:
-        f.write(cmake)
+        if cmake is not None and not os.path.exists(cmake):
+            f.write(cmake)
+        else:
+            f.write(current_cmake)
 
 
 def process_and_move_files(code = None,
